requirements/requirements_mpe.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
astor=0.8.1=pypi_0
astunparse=1.6.3=pypi_0
babel=2.12.1=pypi_0
bce-python-sdk=0.8.87=pypi_0
blinker=1.6.2=pypi_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.4=pypi_0
cloudpickle=1.6.0=pypi_0
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
flask=2.3.2=pypi_0
flask-babel=3.1.0=pypi_0
flask-cors=4.0.0=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.41.0=pypi_0
future=0.18.3=pypi_0
gast=0.4.0=pypi_0
google-auth=2.21.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.56.0=pypi_0
gym=0.23.1=pypi_0
gym-notices=0.0.8=pypi_0
h5py=3.9.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.7.0=pypi_0
importlib-resources=6.0.0=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
keras=2.13.1=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.0=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.9=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pypi_0
paddle-bfloat=0.1.7=pypi_0
paddlepaddle-gpu=2.4.2.post112=pypi_0
pandas=2.0.3=pypi_0
parl=2.2.1=pypi_0
pettingzoo=1.17.0=pypi_0
pillow=10.0.0=pypi_0
pip=23.1.2=py38h06a4308_0
protobuf=3.20.0=pypi_0
psutil=5.9.5=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycryptodome=3.18.0=pypi_0
pyglet=2.0.8=pypi_0
pynvml=11.5.0=pypi_0
pyparsing=3.0.9=pypi_0
python=3.8.17=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
pyzmq=18.1.1=pypi_0
rarfile=4.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.10.1=pypi_0
setuptools=67.8.0=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5=pypi_0
termcolor=2.3.0=pypi_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=pypi_0
visualdl=2.5.3=pypi_0
werkzeug=2.3.6=pypi_0
wheel=0.38.4=py38h06a4308_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h5eee18b_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
